% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{make_MG}
\alias{make_MG}
\title{Create a matched group object}
\usage{
make_MG(unit, MALTS_out)
}
\arguments{
\item{unit}{Query unit whose matched group is desired}

\item{MALTS_out}{An object of type \code{malts}}
}
\value{
An object of type \code{mg.malts}
}
\description{
Create a matched group object
}
