% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{MGs}
\alias{MGs}
\alias{make_MG}
\alias{print.mg.malts}
\alias{summary.mg.malts}
\alias{print.summary.mg.malts}
\alias{plot.mg.malts}
\title{Matched Groups from MALTS}
\usage{
make_MG(unit, MALTS_out, threshold_n, threshold_p)

\method{print}{mg.malts}(x, digits = getOption("digits"), linewidth = 80, ...)

\method{summary}{mg.malts}(object, ...)

\method{print}{summary.mg.malts}(x, digits = getOption("digits"), linewidth = 80, ...)

\method{plot}{mg.malts}(x, cov1, cov2, smooth = FALSE, ...)
}
\arguments{
\item{unit}{Query unit whose matched group is desired}

\item{MALTS_out}{An object of type \code{malts}}

\item{threshold_n, threshold_p}{How often -- across multiple runs and multiple
train-test splits -- a unit must have been matched to another in order to
be considered within the latter's matched group. \code{threshold_p}
specifies a proportion and \code{threshold_n} specifies a number of times.
Defaults to \code{threshold_n = 1}.}

\item{x}{An object of type \code{mg.malts}, returned by a call to \code{MG}.}

\item{digits}{Number of significant digits for printing.}

\item{linewidth}{Maximum number of characters on line; output will be wrapped
accordingly.}

\item{...}{Additional arguments to be passed on to other methods.}

\item{object}{An object of class \code{mg.malts}}

\item{cov1, cov2}{Covariates, specified either by indices or column names in
\code{x$data} to plot against.}

\item{smooth}{A logical scalar denoting whether smoothed loess estimates
should be added to the plots. Defaults to \code{FALSE}.}
}
\value{
\code{make_MG} returns an object of type \code{mg.malts}, which is a
list including the entries:
\itemize{
\item data: treatment, outcome, covariate, and CATE information for the units in
the matched group
\item query: the unit whose matched group is returned
\item threshold: the minimum number of matches required for inclusion in the
matched group. This is the same as \code{threshold_n}, if provided;
otherwise, \code{threshold_p} is converted accordingly.
}

All other entries are as for objects of type \code{malts}


}
\description{
Create, Print, and Plot Matched Groups from a MALTS fit.
}
\details{
\code{print.mg.malts} is equivalent to \code{print.summary.mg.malts}, which
displays information about 1. the size of the query unit's matched group and
2. pruned and unpruned CATE estimates. Unpruned estimates use all the units
the query unit was ever matched to, while pruned estimates only use the units
satisfying the specified thresholds (with equivalence if \code{threshold_n =
1} or \code{threshold_p = 0}).

\code{plot.mg.malts} plots information about the joint distribution of two
covariates within a matched group.
}
\examples{
malts_out <- MALTS(gen_data(n = 500, p = 10), n_folds = 5, n_repeats = 3)
# only including units matched at least 50\% of the time
mg <- make_MG(1, malts_out, threshold_p = 0.5)
print(mg)
plot(mg, 1, 2)

}
