% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{plot.malts}
\alias{plot.malts}
\title{Plot information about a MALTS fit}
\usage{
\method{plot}{malts}(x, which_plots = c(1, 2), ...)
}
\arguments{
\item{x}{An object of class \code{malts}, returned by a call to
\code{\link{MALTS}}.}

\item{which_plots}{A vector describing which plots should be displayed. See
details.}

\item{...}{Additional arguments to passed on to other methods.}
}
\description{
Plot information about the stretch matrix and CATE estimates after a call to
\code{MALTS}.
}
\details{
\code{plot.malts} displays two plots by default. The first shows the diagonal
entries of the stretch matrix \eqn{M} used to define distance between units,
with boxplots over multiple runs (\code{n_repeats}) and train-test splits
(\code{n_folds}). The second plots a density estimate of the estimated CATE
distribution, where the CATEs are (possibly smoothed) averages over multiple
runs (\code{n_repeats}) or train-test splits (\code{n_folds}).
}
