% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{plot.mg.malts}
\alias{plot.mg.malts}
\title{Plot a Matched Group}
\usage{
\method{plot}{mg.malts}(x, cov1, cov2, smooth = FALSE, ...)
}
\arguments{
\item{x}{An object of type \code{mg.malts}, returned by a call to \code{MG}.}

\item{cov1, cov2}{Covariates, specified either by indices or column names in
\code{x$data} to plot against.}

\item{smooth}{A logical scalar denoting whether smoothed loess estimates
should be added to the plots. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to be passed on to other methods.}
}
\description{
Plot information about a matched group after a call to \code{MG}.
}
\details{
\code{plot.mg.malts} displays information about the joint distribution of two
covariates within a matched group.
}
