% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{plot_CATE}
\alias{plot_CATE}
\title{Plot CATEs by Covariate}
\usage{
plot_CATE(malts_out, cov1, condition_on, df, smooth = FALSE)
}
\arguments{
\item{malts_out}{An object of class \code{malts}, returned by a call to
\code{MALTS}.}

\item{cov1}{The variable that CATEs should be plotted against. Either an
index of \code{malts_out$data} corresponding to a covariate or an
appropriate column name.}

\item{condition_on}{A named vector describing what variables to condition on
when plotting CATEs. Names correspond to the variables and how they should
be filtered and values correspond to what values they should be filtered
with respect to. For example c("X1<=" = 5, "X2==" = 0) only plots CATEs for
those units satisfying X1 <= 5 and X2 == 0. Accepted operators are "<",
"<=", ">", ">=", "==", and "!=",  as well as "=" and "", which are
interpreted as "==". See examples.}

\item{df}{A data frame from which CATEs and the corresponding covariate
values are taken to be plotted instead of \code{malts_out$data}. Useful if
more complex filtering than permitted by \code{condition_on} is desired.}

\item{smooth}{A logical scalar, denoting whether a LOESS curve should be fit
to the covariate-CATE data if \code{cov1} is continuous. Defaults to FALSE.}
}
\value{

}
\description{
Plot CATEs by Covariate
}
