% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{print.mg.malts}
\alias{print.mg.malts}
\title{Print a Matched Group}
\usage{
\method{print}{mg.malts}(x, digits = getOption("digits"), linewidth = 80, ...)
}
\arguments{
\item{x}{An object of class \code{mg.malts}.}

\item{digits}{Number of significant digits for printing the CATE estimate.}

\item{linewidth}{Maximum number of characters on line; output will be wrapped
accordingly.}
}
\description{
Print a Matched Group
}
