% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{print.summary.mg.malts}
\alias{print.summary.mg.malts}
\title{Print a Matched Group Summary}
\usage{
\method{print}{summary.mg.malts}(x, digits = getOption("digits"), linewidth = 80, ...)
}
\arguments{
\item{x}{An object of class \code{summary.mg.malts}, generated by a call to
the function of the same name.}

\item{digits}{Number of significant digits for printing.}

\item{linewidth}{Maximum number of characters on line; output will be wrapped
accordingly.}
}
\description{
Print a Matched Group Summary
}
