% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{summary.mg.malts}
\alias{summary.mg.malts}
\title{Summarize a Matched Group}
\usage{
\method{summary}{mg.malts}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mg.malts}}

\item{...}{Additional arguments to be passed on to other methods. Not used.}
}
\value{

}
\description{
Summarize a Matched Group
}
