% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MALTS.R, R/class_funs.R
\name{MALTS}
\alias{MALTS}
\alias{print.malts}
\title{Matching After Learning To Stretch}
\usage{
MALTS(
  data,
  outcome = "outcome",
  treatment = "treated",
  discrete,
  C = 1,
  k_tr = 15,
  k_est = 50,
  reweight = FALSE,
  n_folds = 5,
  n_repeats = 1,
  estimate_CATEs = TRUE,
  smooth_CATEs = TRUE,
  missing_data = c("none", "drop", "impute"),
  impute_with_outcome = FALSE,
  impute_with_treatment = FALSE,
  ...
)

\method{print}{malts}(x, digits = getOption("digits"), linewidth = 80, ...)
}
\arguments{
\item{data}{Data to be matched. Either a data frame or a path to a .csv file
to be read into a data frame. Treatment must be described by a logical or
binary numeric column with name \code{treated_column_name}. If supplied,
outcome must be described by a column with name \code{outcome}.
The outcome will be treated as continuous if numeric with more than two
values, as binary if a two-level factor or numeric with values 0 and 1
exclusively, and as multi-class if a factor with more than two levels. If
the outcome column is omitted, matching will be performed but treatment
effect estimation will not be possible. All columns not containing outcome
or treatment will be treated as covariates for matching.}

\item{outcome}{Name of the outcome column in \code{holdout} and
also in \code{data}, if supplied in the latter. Defaults to 'outcome'.}

\item{treatment}{Name of the treatment column in \code{data} and
\code{holdout}. Defaults to 'treated'.}

\item{discrete}{Specifies which of the covariates in \code{data} are
discrete. If supplied, either a vector of column names or a vector of
column indices corresponding to the discrete covariates. If not supplied,
all covariate columns that are factors will be assumed to be discrete and
the remaining will be assumed to be continuous.}

\item{C}{Regularization weight for the optimization; defaults to 1.}

\item{k_tr, k_est}{Determine matched group sizes for training, matching,
respectively. \code{k_*} many control units \emph{and} \code{k_*} many
treated units are included in each group.}

\item{reweight}{A logical scalar denoting if treated and control groups
should be reweighted during training, according to their respective samples
sizes. Defaults to \code{FALSE}.}

\item{estimate_CATEs}{A logical scalar. If \code{TRUE}, CATEs for each unit
are estimated throughout the matching procedure, which will be much faster
than computing them after a call to \code{MALTS} for very large inputs.
Defaults to \code{TRUE}.}

\item{missing_data}{Specifies how to handle missingness in \code{data}. If
'none' (default), assumes no missing data. If 'drop', effectively drops
units with missingness from the data and does not match them (they will
still appear in the matched dataset that is returned, however). If
'impute', imputes the missing data via \code{mice::mice}.}

\item{impute_with_treatment, impute_with_outcome}{A logical scalar. If
\code{TRUE}, uses treatment, outcome, respectively, to impute covariates
when either \code{missing_data} or \code{missing_holdout} are \code{TRUE}.
Default to \code{TRUE}, \code{FALSE}, respectively. \code{missing_holdout =
  2}. Defaults to \code{TRUE}.}

\item{...}{Additional arguments to be passed to other methods.}

\item{x}{An object of class \code{malts}, returned by a call to
\code{\link{MALTS}}.}

\item{digits}{Number of significant digits for printing the average treatment
effect.}

\item{linewidth}{Maximum number of characters on line; output will be wrapped
accordingly.}

\item{holdout}{Holdout data to be used to estimate a distance matrix. If a
numeric scalar between 0 and 1, that proportion of \code{data} will be made
into a holdout set and only the \emph{remaining proportion} of \code{data}
will be matched. Otherwise, a data frame or a path to a .csv file. The
holdout data must contain an outcome column with name
\code{outcome}; other restrictions on column types are as for
\code{data}. Covariate columns must have the same column names and order as
\code{data}. This data will \emph{not} be matched. Defaults to 0.1.}

\item{repeats}{A positive integer controlling how many times MALTS should be
run. Defaults to 1.}

\item{missing_holdout}{Specifies how to handle missingness in \code{holdout}.
If 'none' (default), assumes no missing data; if 'drop', drops units with
missingness and does not use them to learn a distance matrix; and if
'impute', imputes the missing data via \code{mice::mice}.}
}
\value{
An object of class \code{malts}. The \code{weight} column refers to
the number of matched groups units are included in.
}
\description{
Matching After Learning To Stretch
}
\examples{
malts_out <- MALTS(gen_data(n = 500, p = 10))
}
